requirements.txt
streamlit
praw
pandas
numpy
scikit-learn
matplotlib
joblib
python-dotenv
